restructured the app for easier updating in the future
