+++ Flask_app/oregonwine/forms.py
@@ -1,0 +1,9 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField
+
+class WineFilter(FlaskForm):
+    
+    category = StringField('Wine Category')
+
+
+
